+++ example/android/app/src/main/kotlin/design/codeux/biometric_storage_example/MainActivity.kt
@@ -5,7 +5,7 @@ import io.flutter.embedding.android.FlutterFragmentActivity
 import io.flutter.embedding.engine.FlutterEngine
 
 import io.flutter.plugins.GeneratedPluginRegistrant
-import mu.KotlinLogging
+import io.github.oshai.kotlinlogging.KotlinLogging
 
 private val logger = KotlinLogging.logger {}