+++ macos/Classes/BiometricStorageImpl.swift
@@ -17,10 +17,12 @@ class InitOptions {
     iosTouchIDAuthenticationAllowableReuseDuration = params["iosTouchIDAuthenticationAllowableReuseDurationSeconds"] as? Int
     iosTouchIDAuthenticationForceReuseContextDuration = params["iosTouchIDAuthenticationForceReuseContextDurationSeconds"] as? Int
     authenticationRequired = params["authenticationRequired"] as? Bool
+    darwinBiometricOnly = params["darwinBiometricOnly"] as? Bool
   }
   let iosTouchIDAuthenticationAllowableReuseDuration: Int?
   let iosTouchIDAuthenticationForceReuseContextDuration: Int?
   let authenticationRequired: Bool!
+  let darwinBiometricOnly: Bool!
 }
 
 class IOSPromptInfo {
@@ -138,7 +140,9 @@ class BiometricStorageImpl {
     case .touchIDNotAvailable:
       result("ErrorHwUnavailable")
       break;
-    case .passcodeNotSet: fallthrough
+    case .passcodeNotSet:
+      result("ErrorPasscodeNotSet")
+      break;
     case .touchIDNotEnrolled:
       result("ErrorNoBiometricEnrolled")
       break;
@@ -203,7 +207,7 @@ class BiometricStorageFile {
       guard let access = accessControl(result) else {
         return nil
       }
-      if #available(iOS 13.0, *) {
+      if #available(iOS 13.0, macOS 10.15, *) {
         query[kSecUseDataProtectionKeychain as String] = true
       }
       query[kSecAttrAccessControl as String] = access
@@ -214,10 +218,14 @@ class BiometricStorageFile {
   private func accessControl(_ result: @escaping StorageCallback) -> SecAccessControl? {
     let accessControlFlags: SecAccessControlCreateFlags
     
-    if #available(iOS 11.3, *) {
-      accessControlFlags =  .biometryCurrentSet
+    if initOptions.darwinBiometricOnly {
+      if #available(iOS 11.3, *) {
+        accessControlFlags =  .biometryCurrentSet
+      } else {
+        accessControlFlags = .touchIDCurrentSet
+      }
     } else {
-      accessControlFlags = .touchIDCurrentSet
+      accessControlFlags = .userPresence
     }
         
 //      access = SecAccessControlCreateWithFlags(nil,